pbio/include/pbdrv/cache.h: Turn into proper driver module
This makes this common header extensible in case future platforms
also require cache management. It also makes it clear which functionality
is specific to how Pybricks sets up the EV3.
