+++ lib/pbio/include/pbdrv/cache.h
@@ -3,6 +3,10 @@
 
 #ifndef _PBDRV_CACHE_H_
 
+#include <pbdrv/config.h>
+
+#if PBDRV_CONFIG_CACHE
+
 #include <stddef.h>
 
 // Gets data ready so that the data that we (the CPU) have written
@@ -15,15 +19,17 @@ void pbdrv_cache_prepare_before_dma(const void *buf, size_t sz);
 void pbdrv_cache_prepare_after_dma(const void *buf, size_t sz);
 
 // Accesses a variable via the uncached memory alias
-#define PBDRV_UNCACHED(x)           (*(volatile __typeof__(x) *)((uint32_t)(&(x)) + 0x10000000))
+#define PBDRV_UNCACHED(x)           (*(volatile __typeof__(x) *)((uintptr_t)(&(x)) + PBDRV_CONFIG_CACHE_UNCACHED_OFFSET))
 
 // Gets the address of the uncached memory alias of a variable
-#define PBDRV_UNCACHED_ADDR(x)      ((__typeof__(x) *)((uint32_t)(&(x)) + 0x10000000))
+#define PBDRV_UNCACHED_ADDR(x)      ((__typeof__(x) *)((uintptr_t)(&(x)) + PBDRV_CONFIG_CACHE_UNCACHED_OFFSET))
 
 // Cache line size
 #define PBDRV_CACHE_LINE_SZ         32
 
 // Align data to a cache line, which is needed for clean RX DMA
 #define PBDRV_DMA_BUF               __attribute__((aligned(PBDRV_CACHE_LINE_SZ), section(".dma")))
+
+#endif // PBDRV_CONFIG_CACHE
 
 #endif
